% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nec_code.R
\name{get_all_constituencies}
\alias{get_all_constituencies}
\title{전체 선거구 목록 조회}
\usage{
get_all_constituencies(serviceKey, sgId, sgTypecode)
}
\arguments{
\item{serviceKey}{공공데이터포털에서 발급받은 서비스키}

\item{sgId}{선거ID}

\item{sgTypecode}{선거종류코드}
}
\value{
모든 선거구 데이터프레임
}
\description{
전체 선거구 목록 조회
}
\examples{
get_all_constituencies(serviceKey = "your_service_key", sgId = "20220309", sgTypecode = "1")
}
