% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nec_code.R
\name{get_all_elections}
\alias{get_all_elections}
\title{전체 선거 목록 조회}
\usage{
get_all_elections(serviceKey)
}
\arguments{
\item{serviceKey}{공공데이터포털에서 발급받은 서비스키}
}
\value{
모든 선거 데이터프레임
}
\description{
전체 선거 목록 조회
}
\examples{
get_all_elections(serviceKey = "your_service_key")
}
