% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nec_candidate.R
\name{get_all_preliminary_candidate_info}
\alias{get_all_preliminary_candidate_info}
\title{모든 예비후보자 정보 조회}
\usage{
get_all_preliminary_candidate_info(
  sgId,
  sgTypecode,
  sggName = NULL,
  sdName = NULL
)
}
\arguments{
\item{sgId}{선거ID}

\item{sgTypecode}{선거종류코드}

\item{sggName}{선거구명 (옵션)}

\item{sdName}{시도명 (옵션)}
}
\value{
모든 예비후보자 정보 데이터프레임
}
\description{
모든 예비후보자 정보 조회
}
\examples{
get_all_preliminary_candidate_info(sgId = "20220309", sgTypecode = "1")
}
