% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nec_vote.R
\name{get_all_vote_info}
\alias{get_all_vote_info}
\title{모든 투표결과 조회}
\usage{
get_all_vote_info(serviceKey, sgId, sgTypecode, sdName = NULL, wiwName = NULL)
}
\arguments{
\item{serviceKey}{공공데이터포털에서 발급받은 서비스키}

\item{sgId}{선거ID}

\item{sgTypecode}{선거종류코드}

\item{sdName}{시도명 (옵션)}

\item{wiwName}{구시군명 (옵션)}
}
\value{
모든 투표결과 데이터프레임
}
\description{
모든 투표결과 조회
}
\examples{
get_all_vote_info(serviceKey = "your_service_key", sgId = "20220309", sgTypecode = "1")
}
