% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nec_constituency_voters.R
\name{get_all_voters}
\alias{get_all_voters}
\title{모든 선거인수 정보 조회}
\usage{
get_all_voters(sgId, sgTypecode = NULL, sdName = NULL, wiwName = NULL)
}
\arguments{
\item{sgId}{선거ID}

\item{sgTypecode}{선거종류코드 (선거구별 조회 시 필요)}

\item{sdName}{시도명 (옵션)}

\item{wiwName}{구시군명 (옵션)}
}
\value{
모든 선거인수 정보를 포함하는 리스트
}
\description{
모든 선거인수 정보 조회
}
\examples{
get_all_voters(sgId = "20200415", sgTypecode = "2", sdName = "서울특별시")
}
