% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nec_winner.R
\name{get_all_winner_info}
\alias{get_all_winner_info}
\title{전체 당선인 정보 조회}
\usage{
get_all_winner_info(sgId, sgTypecode, sdName = NULL, sggName = NULL)
}
\arguments{
\item{sgId}{선거ID}

\item{sgTypecode}{선거종류코드}

\item{sdName}{시도명 (옵션)}

\item{sggName}{선거구명 (옵션)}
}
\value{
전체 당선인 정보 데이터프레임
}
\description{
선거ID, 선거종류코드, 시도명, 선거구명을 입력받아 모든 페이지의 당선인 정보를 조회하는 함수
}
\examples{
get_all_winner_info(sgId = "20220309", sgTypecode = "1")
}
