% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nec_code.R
\name{get_constituencies}
\alias{get_constituencies}
\title{선거구 목록 조회 (단일 페이지)}
\usage{
get_constituencies(serviceKey, pageNo = 1, numOfRows = 10, sgId, sgTypecode)
}
\arguments{
\item{serviceKey}{공공데이터포털에서 발급받은 서비스키}

\item{pageNo}{페이지 번호 (기본값: 1)}

\item{numOfRows}{페이지당 데이터 수 (기본값: 10)}

\item{sgId}{선거ID}

\item{sgTypecode}{선거종류코드}
}
\value{
선거구 데이터프레임
}
\description{
선거구 목록 조회 (단일 페이지)
}
\examples{
get_constituencies(serviceKey = "your_service_key", sgId = "20220309", sgTypecode = "1")
}
