% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nec_constituency_voters.R
\name{get_constituency_voters}
\alias{get_constituency_voters}
\title{선거구별 선거인수 정보 조회}
\usage{
get_constituency_voters(
  sgId,
  sgTypecode,
  sdName = NULL,
  wiwName = NULL,
  pageNo = 1,
  numOfRows = 10
)
}
\arguments{
\item{sgId}{선거ID}

\item{sgTypecode}{선거종류코드}

\item{sdName}{시도명 (옵션)}

\item{wiwName}{구시군명 (옵션)}

\item{pageNo}{페이지 번호 (기본값: 1)}

\item{numOfRows}{페이지당 출력 개수 (기본값: 10, 최대값: 100)}
}
\value{
선거구별 선거인수 정보 데이터프레임
}
\description{
선거구별 선거인수 정보 조회
}
\examples{
get_constituency_voters(sgId = "20200415", sgTypecode = "2", sdName = "서울특별시")
}
