% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nec_election_code.R
\name{get_occupations}
\alias{get_occupations}
\title{직업 목록 조회 (단일 페이지)}
\usage{
get_occupations(pageNo = 1, numOfRows = 10, sgId)
}
\arguments{
\item{pageNo}{페이지 번호 (기본값: 1)}

\item{numOfRows}{페이지당 데이터 수 (기본값: 10)}

\item{sgId}{선거ID}
}
\value{
직업 데이터프레임
}
\description{
직업 목록 조회 (단일 페이지)
}
\examples{
get_occupations(sgId = "20220309")
}
