% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nec_candidate.R
\name{get_preliminary_candidate_info}
\alias{get_preliminary_candidate_info}
\title{예비후보자 정보 조회}
\usage{
get_preliminary_candidate_info(
  serviceKey,
  sgId,
  sgTypecode,
  sggName = NULL,
  sdName = NULL,
  pageNo = 1,
  numOfRows = 10
)
}
\arguments{
\item{serviceKey}{공공데이터포털에서 발급받은 서비스키}

\item{sgId}{선거ID}

\item{sgTypecode}{선거종류코드}

\item{sggName}{선거구명 (옵션)}

\item{sdName}{시도명 (옵션)}

\item{pageNo}{페이지 번호 (기본값: 1)}

\item{numOfRows}{페이지당 출력 개수 (기본값: 10, 최대값: 100)}
}
\value{
예비후보자 정보 데이터프레임
}
\description{
예비후보자 정보 조회
}
\examples{
get_preliminary_candidate_info(serviceKey = "your_service_key", sgId = "20220309", sgTypecode = "1")
}
