% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nec_constituency_voters.R
\name{get_sido_voters}
\alias{get_sido_voters}
\title{시도별 선거인수 정보 조회}
\usage{
get_sido_voters(sgId, pageNo = 1, numOfRows = 10)
}
\arguments{
\item{sgId}{선거ID}

\item{pageNo}{페이지 번호 (기본값: 1)}

\item{numOfRows}{페이지당 출력 개수 (기본값: 10, 최대값: 100)}
}
\value{
시도별 선거인수 정보 데이터프레임
}
\description{
시도별 선거인수 정보 조회
}
\examples{
get_sido_voters(sgId = "20200415")
}
