% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nec_election_code.R
\name{get_sigungu}
\alias{get_sigungu}
\title{행정구역 목록 조회 (단일 페이지)}
\usage{
get_sigungu(pageNo = 1, numOfRows = 10, sgId, sdName = NULL)
}
\arguments{
\item{pageNo}{페이지 번호 (기본값: 1)}

\item{numOfRows}{페이지당 데이터 수 (기본값: 10)}

\item{sgId}{선거ID}

\item{sdName}{시도명 (기본값: NULL)}
}
\value{
행정구역 데이터프레임
}
\description{
행정구역 목록 조회 (단일 페이지)
}
\examples{
get_sigungu(sgId = "20220309")
}
