% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nec_winner.R
\name{get_winner_info}
\alias{get_winner_info}
\title{당선인 정보 조회}
\usage{
get_winner_info(
  sgId,
  sgTypecode,
  sdName = NULL,
  sggName = NULL,
  pageNo = 1,
  numOfRows = 10,
  resultType = "json"
)
}
\arguments{
\item{sgId}{선거ID}

\item{sgTypecode}{선거종류코드}

\item{sdName}{시도명 (옵션)}

\item{sggName}{선거구명 (옵션)}

\item{pageNo}{페이지 번호 (기본값: 1)}

\item{numOfRows}{목록 건수 (기본값: 10, 최대값: 100)}

\item{resultType}{데이터포맷 (기본값: "json", 다른 옵션: "xml")}
}
\value{
당선인 정보 데이터프레임
}
\description{
선거ID, 선거종류코드, 시도명, 선거구명을 입력받아 당선인 관련 정보를 조회하는 함수
}
\examples{
get_winner_info(sgId = "20220309", sgTypecode = "1")
}
