% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nec_winner.R
\name{getAllWinnerInfo}
\alias{getAllWinnerInfo}
\title{전체 당선인 정보 조회}
\usage{
getAllWinnerInfo(serviceKey, sgId, sgTypecode, sdName = NULL, sggName = NULL)
}
\arguments{
\item{serviceKey}{공공데이터포털에서 발급받은 서비스키}

\item{sgId}{선거ID}

\item{sgTypecode}{선거종류코드}

\item{sdName}{시도명 (옵션)}

\item{sggName}{선거구명 (옵션)}
}
\value{
전체 당선인 정보 데이터프레임
}
\description{
선거ID, 선거종류코드, 시도명, 선거구명을 입력받아 모든 페이지의 당선인 정보를 조회하는 함수
}
\examples{
getAllWinnerInfo(serviceKey = "your_service_key", sgId = "20220309", sgTypecode = "1")
}
