% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nec_code.R
\name{getCommonEduBckgrdCodeList}
\alias{getCommonEduBckgrdCodeList}
\title{학력코드 조회}
\usage{
getCommonEduBckgrdCodeList(serviceKey, pageNo = 1, numOfRows = 10, sgId)
}
\arguments{
\item{serviceKey}{공공데이터포털에서 발급받은 서비스키}

\item{pageNo}{페이지 번호 (기본값: 1)}

\item{numOfRows}{페이지당 데이터 수 (기본값: 10)}

\item{sgId}{선거ID}
}
\value{
학력코드 데이터프레임
}
\description{
선거ID를 입력하여 학력코드, 학력명, 순서 데이터를 조회하는 함수
}
\examples{
getCommonEduBckgrdCodeList(serviceKey = "your_service_key", sgId = "20220309")
}
