% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nec_code.R
\name{getCommonGusigunCodeList}
\alias{getCommonGusigunCodeList}
\title{구시군코드 조회}
\usage{
getCommonGusigunCodeList(
  serviceKey,
  pageNo = 1,
  numOfRows = 10,
  sgId,
  sdName = NULL
)
}
\arguments{
\item{serviceKey}{공공데이터포털에서 발급받은 서비스키}

\item{pageNo}{페이지 번호 (기본값: 1)}

\item{numOfRows}{페이지당 데이터 수 (기본값: 10)}

\item{sgId}{선거ID}

\item{sdName}{시도명 (기본값: NULL)}
}
\value{
구시군코드 데이터프레임
}
\description{
선거ID를 입력하여 구시군명, 순서, 상위시도명 데이터를 조회하는 함수
}
\examples{
getCommonGusigunCodeList(serviceKey = "your_service_key", sgId = "20220309")
}
