% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nec_code.R
\name{getCommonSgCodeList}
\alias{getCommonSgCodeList}
\title{선거코드 조회}
\usage{
getCommonSgCodeList(serviceKey, pageNo = 1, numOfRows = 10)
}
\arguments{
\item{serviceKey}{공공데이터포털에서 발급받은 서비스키}

\item{pageNo}{페이지 번호 (기본값: 1)}

\item{numOfRows}{페이지당 데이터 수 (기본값: 10)}
}
\value{
선거코드 데이터프레임
}
\description{
선거코드, 선거명, 선거종류코드, 선거일자 데이터를 조회하는 함수
}
\examples{
getCommonSgCodeList(serviceKey = "your_service_key")
}
