% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nec_code.R
\name{getCommonSggCodeList}
\alias{getCommonSggCodeList}
\title{선거구코드 조회}
\usage{
getCommonSggCodeList(serviceKey, pageNo = 1, numOfRows = 10, sgId, sgTypecode)
}
\arguments{
\item{serviceKey}{공공데이터포털에서 발급받은 서비스키}

\item{pageNo}{페이지 번호 (기본값: 1)}

\item{numOfRows}{페이지당 데이터 수 (기본값: 10)}

\item{sgId}{선거ID}

\item{sgTypecode}{선거종류코드}
}
\value{
선거구코드 데이터프레임
}
\description{
선거ID와 선거종류코드를 입력하여 선거구명, 시도명, 구시군명, 선출정수, 순서 데이터를 조회하는 함수
}
\examples{
getCommonSggCodeList(serviceKey = "your_service_key", sgId = "20220309", sgTypecode = "1")
}
