% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nec_election_code.R
\name{get_all_sigungu}
\alias{get_all_sigungu}
\title{전체 행정구역 목록 조회}
\usage{
get_all_sigungu(sgId, sdName = NULL)
}
\arguments{
\item{sgId}{선거ID}

\item{sdName}{시도명 (기본값: NULL)}
}
\value{
모든 행정구역 데이터프레임
}
\description{
전체 행정구역 목록 조회
}
\examples{
get_all_sigungu(sgId = "20220309")
}
